# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dropbox PHP",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/trash": {
            "post": {
                "tags": [
                    "Trash"
                ],
                "summary": "Trash structure display",
                "description": "Send bearer token",
                "operationId": "6cc79a09f632e9fc54f5793ee2ec6f08",
                "responses": {
                    "200": {
                        "description": "Trash structure display",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "trash_lifespan": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "folders": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "folder_name": {
                                                            "type": "string",
                                                            "example": "Holidays"
                                                        },
                                                        "created_by_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "files": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "name": {
                                                                "type": "string",
                                                                "example": "IMG_0514"
                                                            },
                                                            "file_name": {
                                                                "type": "string",
                                                                "example": "IMG_0514.jpg"
                                                            },
                                                            "uuid": {
                                                                "type": "string",
                                                                "example": "5a3e86e4-c09d-4594-8bf4-be8776e8769f"
                                                            },
                                                            "preview_url": {
                                                                "type": "string",
                                                                "example": ""
                                                            },
                                                            "original_url": {
                                                                "type": "string",
                                                                "example": "http://localhost/storage/10/IMG_0514.JPG"
                                                            },
                                                            "extension": {
                                                                "type": "string",
                                                                "example": "JPG"
                                                            },
                                                            "size": {
                                                                "type": "integer",
                                                                "example": 5199684
                                                            },
                                                            "folder_id": {
                                                                "type": "integer",
                                                                "example": 4
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/trash/delete": {
            "delete": {
                "tags": [
                    "Trash"
                ],
                "summary": "Removing cart items",
                "description": "Send bearer token, and element ids. key-value bindings can be repeated",
                "operationId": "0716129c63a81c5162676dce98ce72cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "folder": {
                                        "type": "string",
                                        "example": "3"
                                    },
                                    "file": {
                                        "type": "string",
                                        "example": "20"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Removing cart items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "All items removed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    }
}